requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
RPi.GPIO==0.7.1
websockets==12.0
python-multipart==0.0.6
jinja2==3.1.2
aiofiles==23.2.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0